Extract exploration column components
